# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.APIError:
    properties:
      details: {}
      message:
        type: string
    type: object
...
  handler.CategoryResponse:
    properties:
      active:
        type: boolean
      colorHex:
        type: string
      createdAt:
        type: string
      icon:
        type: string
      id:
        type: string
      name:
        type: string
      order:
        type: integer
      type:
        type: string
      updatedAt:
        type: string
    type: object
...
  handler.ExpenseResponse:
    properties:
      amount:
        type: number
      category:
        $ref: '#/definitions/handler.CategoryResponse'
      categoryId:
        type: string
      createdAt:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      origin:
        type: string
      receipt:
        $ref: '#/definitions/handler.ReceiptResponse'
      recurring:
        type: boolean
      updatedAt:
        type: string
    type: object
...
  handler.ReceiptItem:
    properties:
      description:
        type: string
      quantity:
        type: number
      total:
        type: number
      unitPrice:
        type: number
    type: object
  handler.ReceiptResponse:
    properties:
      extractedText:
        type: string
      filePath:
        type: string
      id:
        type: string
      ocrConfidence:
        type: number
    type: object
  handler.ReceiptScanRequest:
    properties:
      amountHint:
        type: number
      currency:
        type: string
      imageBase64:
        type: string
      locale:
        type: string
      returnRaw:
        type: boolean
    type: object
  handler.ReceiptScanResponse:
    properties:
      confidence:
        type: number
      currency:
        type: string
      extractedText:
        type: string
      items:
        items:
          $ref: '#/definitions/handler.ReceiptItem'
        type: array
      model:
        type: string
      rawModelOutput:
        type: string
      savedExpense:
        $ref: '#/definitions/handler.ExpenseResponse'
      suggestedAmount:
        type: number
      suggestedDate:
        type: string
      tokenCostCents:
        type: integer
      tokensUsed:
        type: integer
    type: object
...
  handler.SyncJobResponse:
    properties:
      finishedAt:
        type: string
      id:
        type: string
      origin:
        type: string
      startedAt:
        type: string
      status:
        type: string
    type: object
  handler.SyncRequest:
    properties:
      origin:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: contato@example.com
    name: Equipe Golang Finance
  description: API para gestão financeira pessoal com suporte a OCR de recibos.
  title: Golang Finance API
  version: "1.0"
paths:
...
  /receipts/scan:
    post:
      consumes:
      - application/json
      description: Analisa uma imagem em Base64 usando Gemini e retorna extrações
        estruturadas
      parameters:
      - description: Dados do recibo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.ReceiptScanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ReceiptScanResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Processar recibo com OCR
      tags:
      - Recibos
  /sync/jobs:
    post:
      consumes:
      - application/json
      description: Salva o histórico de uma sincronização concluída
      parameters:
      - description: Origens
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.SyncRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SyncJobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Registrar sincronização
      tags:
      - Sincronização
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"